requirements_txt.txt
# MCP Network Upgrader Requirements

# MCP Framework
mcp>=1.0.0

# AI/LLM Integration
anthropic>=0.27.0
openai>=1.0.0

# Database Clients
psycopg2-binary>=2.9.0
influxdb-client>=1.38.0
redis>=4.5.0

# Ansible and Network
ansible>=8.0.0
paramiko>=3.0.0
netmiko>=4.2.0

# Web Interface
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Telegram Bot
python-telegram-bot>=20.0

# Data Validation and Config
pydantic>=2.4.0
pyyaml>=6.0
python-dotenv>=1.0.0

# CLI and Utilities
click>=8.1.0
rich>=13.5.0
tabulate>=0.9.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0